fix: cast crashes to unsigned long to match %lu format specifier
GCC -Wformat=2 -Werror: uint32_t (unsigned int) != unsigned long even
when both are 32-bit on Xtensa.

+++ src/boot.cpp
@@ -33,7 +33,7 @@ void safeBootCheck() {
         }
         esp_deep_sleep_start();
     } else {
-        LOG_WARN("crash %lu/3 — incrementing counter", crashes);
+        LOG_WARN("crash %lu/3 — incrementing counter", (unsigned long)crashes);
         p.putUInt("crashes", crashes);
         p.end();
     }
